Flatten article body into a single rich text field
